% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Get prior or posterior predictive samples for the model}
\usage{
run_model(n, diff, age, age2, intercept.mu, intercept.variance, age.mu,
  age.variance, age2.mu, age2.variance, kappa0, theta0, run_estimation)
}
\arguments{
\item{n}{number of observations}

\item{diff}{outcome variable}

\item{age}{linear predictor}

\item{age2}{quadratic predictor}

\item{intercept.mu}{hyperparameter mean intercept prior}

\item{intercept.variance}{hyperparameter variance intercept prior, note not sd}

\item{age.mu}{hyperparameter mean age prior}

\item{age.variance}{hyperparameter variance age prior, note not sd}

\item{age2.mu}{hyperparameter mean age2 prior}

\item{age2.variance}{hyperparameter variance age2 prior, note not sd}

\item{kappa0}{shape inverse gamma prior on residual variance, note not sd}

\item{theta0}{rate inverse gamma prior on residual variance, note not sd}

\item{run_estimation}{to evaluate likelihood or not, thus 0 to get prior predictive 1 to get posterior predictive}
}
\value{
returns prior of posterior predictive samples.
}
\description{
Enter a more detailed description if you like.
}
\examples{
\dontrun{
run_model(10, rnorm(10), rnorm(10), rnorm(10), 0, 1, 0, 1, 0, 1, .5, .5, 0)
}

}
