% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackages-package.R
\docType{package}
\name{testpackages-package}
\alias{testpackages-package}
\alias{testpackages}
\title{The 'testpackages' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
